requirements.txt
alembic==1.7.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
flake8==4.0.1
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.36
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
